ARM: tegra: mc: introduce io wrappers
Instead of using readl() and writel() directly introduce an
API for reading and writing to the MC register spaces. The API
is as follows:

  u32  mc_readl(u32 reg_offset);
  void mc_writel(u32 val, u32 reg_offset);
  u32  __mc_readl(int mc_idx, u32 reg_offset);
  void __mc_writel(int mc_idx, u32 val, u32 reg_offset);

The default functions read and write from MC0 - that is the
interface all chips have. However, some chips have dual
channel MCs. The '__' versions allow you to specify which MC
to read/write to/from.

Change-Id: I352c0f42ebe008b11586e15ce6b73bd46d7368b5
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Reviewed-on: http://git-master/r/302570
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>