video: tegra: host: gk20a: Imm TLB invalidation
PMU assumes that the pages that it requests to be mapped are available
immediately. However, we currently postpone TLB invalidation to channel
submission and therefore the hw may end up accessing the pages
before the TLB contains the pages.

This patch adds TLB invalidation when we are mapping memory to non-
existent address space (i.e. pmu and fecs memory).

Bug 1334322

Change-Id: I1ed48d224589474a0cd438076478ff1c6ffecc5a
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/300583
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>