video: tegra: host: Fix deadlock in rl update
Engine recovery was initiated if the runlist update failed. The
recovery is performed in asynchronous fashion and it accesses runlist
data structures that are protected by runlist lock. Runlist update routine,
however, did not release lock at this point and therefore it effectively
caused deadlock.

This patch makes the runlist routine to release the mutex if it needs to
initiate engine recovery

Change-Id: Ifa990117fa34a31cff73ce3ea64e67720f28aede
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/300625
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>
Tested-by: Terje Bergstrom <tbergstrom@nvidia.com>