video: tegra: dc: no HPD probing in atomic context
tegra_dc_hpd() uses a potentially sleeping gpio_get() call to check HPD
status. Move the call to tegra_dc_hpd() out of the atomic context to
prevent a potential lock.

The spinlock we are getting away from is supposed to protect the
worker structure and has nothing to do with HPD itself. By probing
HPD status once the worker structure has been updated, we are sure to
not miss any HPD event.

Change-Id: If80a49d82355aa24570bd43c5101f02588f6468f
Signed-off-by: Alexandre Courbot <acourbot@nvidia.com>
Signed-off-by: Xia Yang <xiay@nvidia.com>
Reviewed-on: http://git-master/r/354013
Reviewed-by: Automatic_Commit_Validation_User
Tested-by: Jon Mayo <jmayo@nvidia.com>
Reviewed-by: Jon Mayo <jmayo@nvidia.com>