usb: gadget: tcm: Fix UASP enumeration issue
Currently the same endpoint descriptor
is used in the configuration descriptor
(uasp_ss_function_desc) for  both the BOT and
UASP. So in the init when we call the
usb_ep_autoconfig_ss, the address is set in the
uasp_ss_*_desc common descriptor. Later when we
try to configure the ep by calling
config_ep_by_speed,it matches the command ep
descriptor of BOT in the configuration descriptor
and uses the next bot’s companion descriptor that
doesn’t have the streams enabled. Because of
using this wrong descriptor the UASP is not
getting enumerated.

Change-Id: I72da7ed13b738498b2427bfe9a4e4e101d49cddf
Signed-off-by: Ajay Gupta <ajayg@nvidia.com>
Signed-off-by: Hui Fu <hfu@nvidia.com>
Reviewed-on: http://git-master/r/439409
Reviewed-by: Ashutosh Jha <ajha@nvidia.com>