arm64: fix access to preempt_count from assembly code
preempt_count is defined as an int. Oddly enough, we access it
as a 64bit value. Things become interesting when running a BE
kernel, and looking at the current CPU number, which is stored
as an int next to preempt_count. Like in a per-cpu interrupt
handler, for example...

Change-Id: I991043104e859e40e903f264943b23345f817baa

Using a 32bit access fixes the issue for good.
(cherry picked from commit 717321fcb58ed95169bf344ae47ac6098ba5dfbe)
Cc: Matthew Leach <matthew.leach@arm.com>
Signed-off-by: Marc Zyngier <Marc.Zyngier@arm.com>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>