video: tegra: host: gk20a: free irq for railgate
There is a possibility of spurious interrupts being generated.
If we attempt to service these when gk20a is railgated, they will
almost certainly hang the system. Hence we free the irq just before
suspending, to avoid the handler being called. Also, we re-request
the irq at the beginning of finalize_poweron.

Bug 1404859
Bug 1364240
Bug 1392195

Change-Id: Ic54c72886d91dcc50c51cbf1565e77ab08c55cec
Signed-off-by: Prashant Malani <pmalani@nvidia.com>
Reviewed-on: http://git-master/r/327958
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>
Tested-by: Terje Bergstrom <tbergstrom@nvidia.com>