Revert "ARM: arch_timer: use full 64-bit counter for sched_clock"
This reverts commit 023796b9be3a77481cd5ee0b64a13a55bb0d5df4.

With reverting this change we are moving back to the
32 bit counter instead of the 56 bit counter.

The change is reverted because the multiplier in this
case is 80 times the original counter due to which although
the sched clock is monotonic the time read by clients
are further ahead of the actual time.

This Change will be restored once we start taking the upstream
patches.

Bug 1399318

Signed-off-by: Ajay Nandakumar <anandakumarm@nvidia.com>
Change-Id: I5aebc5fe2a1096d5fa2acee7edcd7e03f3c7c565
Reviewed-on: http://git-master/r/328642
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>