video: tegra: host: pod: Init last req correctly
The scaling initialisation read the current frequency and stored
it as the "last requested frequency" into devfreq data structures.
If the gpu was disabled at this point, this operation initialised
previous_frequency to 0. The scaling code uses the last requested
frequency and load to determine optimal frequency. If the data
was not available, we skipped scaling.

This code modifies the scaling so that:
1) we never touch previous_frequency. devfreq does that
2) we give the minimum frequency to devfreq as a parameter. devfreq
can then use that value while initialising previous_frequency.

Bug 1418949

Change-Id: I13c834735d68d8ff7441a77b0aad6c567bb82f40
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/339880
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Ilan Aelion <iaelion@nvidia.com>
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>