video: tegra: host: Fixes to PMU perfmon code
PMU bases timings on the clockrate that is given inside an
initialisation message. If this rate is wrong, we do perfmon
calculations incorrectly. This far we have assumed fixed 500MHz clock
which is not correct.

This patch modifies the PMU initialisation to actually read the
PMU clock frequency from the clock framework. In addition, this patch
modifies perfmon to capture samples 10 times per frame.

Bug 1422029

Change-Id: I7d700a4eeba43630e4f9332bb997cccf19e72aff
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/343562
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>
Tested-by: Terje Bergstrom <tbergstrom@nvidia.com>