ARM: dma-mapping: alloc highmem w/o GFP_DMA*
There's a case that a client wants to avoid highmem explicitly. For
example some PM code lets firmware access to pages during
suspend/resume when IOMMU is off. This allows a client to specify a
type of pages{GPF_DMA,GFP_DMA32} to not contradicting to
GFP_HIGHMEM. This would be valid if AArch64 selects ZONE_DMA(32).

Bug 1414172

Change-Id: I3c97262f9388094dae12a350420a3d184d1e7144
Signed-off-by: Hiroshi Doyu <hdoyu@nvidia.com>
Reviewed-on: http://git-master/r/343645
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>
Tested-by: Krishna Reddy <vdumpa@nvidia.com>