arch: arm: tegra: vcm30t124: Make LDO5 always on
LDO5 is used to control sdmmc voltage, which is either 1.8V or 3.3V. And
since it controlled via FPS1 of MAX77663, we can't disable/enable the
rail from the S/w unless we move the rail out of FPS control.
So make the LDO5 always on in the S/w, so that regulator framework
doesn't try to call enable/disable routine.

bug 1397523

Change-Id: Ice3d415341a70092835730fddfd106cd8786fba0
Signed-off-by: Bhavesh Parekh <bparekh@nvidia.com>
Reviewed-on: http://git-master/r/344726
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Sumeet Gupta <sumeetg@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>