drivers: platform: tegra: add adsp driver
The Audio DSP (ADSP) is used for audio specific processing. The host
needs to communicate with ADSP for various operations. This ADSP
driver provides a set of APIs which can be used for managing these
operations.

Currently, only stubs are added in the driver. These will be
implemented one by one to enable particular features.

Bug 1445394

Change-Id: I21f401ab5ce0a7ce70e9f6437df4e36aa0a08a0b
Signed-off-by: Nitin Kumbhar <nkumbhar@nvidia.com>
Reviewed-on: http://git-master/r/350421
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: Bo Yan <byan@nvidia.com>