video: tegra: host: Add waits to FECS reset
Based on discussion with hw folks, the correct way to wait reset
is to first read back the reset register and then wait a moment. This
should be done also after we release the reset to ensure that the hw
is ready before we actually do anything with it.

Currently the code just has a predefined wait which may cause unexpected
behaviour of the hw.

Bug 1425077

Change-Id: I477350df04f2045da84a6084db0d634df2bf617c
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/350914
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>
Tested-by: Terje Bergstrom <tbergstrom@nvidia.com>