arm: tegra: loki: Support to read battery temp through PMIC
This change enables reading the battery temperature through
PMIC ADC. The adc value conversion to temperature is performed
using a look-up table that maps the adc values to temp. The
conversion table is based on values generated for Loki
P2530 board.

More details on the source of the table can be
found in the bug.

Bug 1408758

Change-Id: I24aa6eb4f23f9fb465da2e28105da5ecdd113683
Signed-off-by: Ankita Garg <ankitag@nvidia.com>
Reviewed-on: http://git-master/r/356182
Reviewed-by: Diwakar Tundlam <dtundlam@nvidia.com>
Tested-by: Diwakar Tundlam <dtundlam@nvidia.com>