arm: tegra: loki: Batt temp and tdiode shutdown
This change enables reading the battery temperature and
Tdiode through PMIC ADC and shutdown based on that.
The adc value conversion to temperature
is performed using a look-up table that maps the adc values
to temp. The conversion table is based on values generated
for Loki P2530 board.

Bug 1408758
Bug 1398960

Change-Id: Ic27efe4f7b608ee355799c9498413a862dd4ea39
Signed-off-by: Ankita Garg <ankitag@nvidia.com>
Signed-off-by: Anshul Jain <anshulj@nvidia.com>
Reviewed-on: http://git-master/r/375907
Reviewed-by: Diwakar Tundlam <dtundlam@nvidia.com>
Tested-by: Diwakar Tundlam <dtundlam@nvidia.com>