ARM: tegra: dvfs: Parse DFLL device tree data
Added parsing of DFLL device tree data in tegra CL-DVFS driver probe.
Implementation is backward compatible with the specification of DFLL
platform data in board files - if the latter is present, DT data is
ignored.

Only PMIC with I2C interface can be integrated with DFLL data in DT.
For PMIC with PWM interface integration data must still be specified
as static platform data in the respective board file.

Bug 1442709

Change-Id: Ie481c0f766b13308449199168e3567711a8834f3
Signed-off-by: Alex Frid <afrid@nvidia.com>
Reviewed-on: http://git-master/r/357461
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Seshendra Gadagottu <sgadagottu@nvidia.com>
Reviewed-by: Bo Yan <byan@nvidia.com>