arm64: create an option to use simple spinlocks
The ticket based spin locks may reduce the worst case latency  in
use cases where there is a lot of contention for a lock. However,
they require more overhead than a simple lock in the case that there
is no contention.

This implements a simple spinlock which improves performance in the
use cases where there is not much contention and fairness is not
a big problem.

Also, the implementation of the ticket based lock for the arm64
architecture does not use the WFE instruction optimally. It uses the
idiom:
	sevl
1:	wfe
	ldaxr
	cmp
	bne 1b

This is not ideal in the contested or uncontested case because it
delays executing the ldaxr instruction. It executes potentially slow
WFE and SEVL instructions.

Further, this idiom is incompatible with __raw_spin_lock which uses:

while (!raw_spin_can_lock(lock) && (lock)->break_lock)
arch_spin_relax(&lock->raw_lock);

The problem is that there is no place to put the SEVL outside the
loop. This results in the CPU busy spinning anytime there is
contention for a spinlock.

A better idiom in all cases is something like:

1:	ldaxr
	cmp
	bne 2f
	wfe
	b 1b
2:

With this idiom, the load is executed immediately instead of executing
the potentially slow no-op SEVL+WFE.

This idiom fits nicely into the __raw_spin_lock loop allowing the CPU
to go to sleep while waiting for the lock.

Bug 1440840

Change-Id: Ic048b84bae8a71513f697e4e9a9eae7b6f106f54
Signed-off-by: Alex Van Brunt <avanbrunt@nvidia.com>
Reviewed-on: http://git-master/r/369018