drivers: phy: Make NULL a valid phy reference
The common clock framework considers NULL a valid clock
reference. This makes handling optional clocks simple, in that if the
optional clock is not available, a NULL reference can be used in the
place of a real clock, simplifying the clock consumer.

Extend this concept to the phy consumer API. A NULL can be passed to
the release calls, the phy_init() and phy_exit() calls, and
phy_power_on() and phy_power_off() and a NOP is performed.

Signed-off-by: Andrew Lunn <andrew@lunn.ch>
Signed-off-by: Jason Cooper <jason@lakedaemon.net>
(cherry picked from commit 04c2facad8fee66c981a51852806d8923336f362)

Bug 200024489

Signed-off-by: Rakesh Babu Bodla <rbodla@nvidia.com>
Change-Id: I4469be993728e02766413b0af466c1e1e690bda5
Reviewed-on: http://git-master/r/682486
GVS: Gerrit_Virtual_Submit
Reviewed-by: Venkat Moganty <vmoganty@nvidia.com>