ARM64: perf: support dwarf unwinding in compat mode
Add support for unwinding using the dwarf information in compat
mode. Using the correct user stack pointer allows perf to record
the frames correctly in the native and compat modes.

Note that although the dwarf frame unwinding works ok using
libunwind in native mode (on ARMv7 & ARMv8), some changes are
required to the libunwind code for the compat mode. Those changes
are posted separately on the libunwind mailing list.

Tested on ARMv8 platform with v8 and compat v7 binaries, the latter
are statically built.

Change-Id: I10be6930346f998ed1facb70c54b2ff39e59d559
Signed-off-by: Jean Pihet <jean.pihet@linaro.org>
Acked-by: Will Deacon <will.deacon@arm.com>
Reviewed-on: http://git-master/r/379164
Reviewed-by: Ryan Bissell <rbissell@nvidia.com>
Tested-by: Ryan Bissell <rbissell@nvidia.com>
Reviewed-by: Thomas Cherry <tcherry@nvidia.com>