video: tegra: host: Remove ext_stat from scaling
This far we have delivered minimum frequency, maximum frequency and
the last event type in private_data field for the governor. All these
fields are already available in some other format so it makes no
sense to duplicate this information.

This patch makes necessary modifications to remove dependency to
ext_stat structure.

Bug 1454499

Change-Id: I076aaeb0b9204b810f345ab03b648e238cdab3ff
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/365422
Reviewed-by: Hiroshi Doyu <hdoyu@nvidia.com>
Tested-by: Hiroshi Doyu <hdoyu@nvidia.com>