ARM: tegra: remove ddr rail voltage setting explictly for E1922
E1922 uses the E1936-1100 PMIC module. This PMIC module has
ddr rail voltage of 1.2V as default and hence it is not require
to override this value.

E1936-1000 is used on system with DDR3 memory where ddr rail voltage
is set to 1.35V.

So in both cases, the PMIC default is as per the memory requirements
and hence it is not require to set the voltage explictly.

Change-Id: I31ae993dfb07d61902d50e35e62872d9cb842450
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/366569