driver: media: tegra: Update the power state.
bug 1457201

Update the power state when powe on. Or it will cause power off failure.
For instance, when the first time invoking ov5693_pm_wr to set the pwr_dev to 2,
the second time to power off ov5693 will be failure if the ov5693 is already on
but not updating the power state.

Change-Id: I5f307b302d7f4d7afae6e900a84103f4b36e2605
Signed-off-by: Jiukai Ma <jiukaim@nvidia.com>
Reviewed-on: http://git-master/r/367196
(cherry picked from commit cce945356e91b5c74eb82644a6277932f5074413)
Reviewed-on: http://git-master/r/375513
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Frank Chen <frankc@nvidia.com>
Reviewed-by: Mitch Luban <mluban@nvidia.com>